+++ terraform/pagerduty/terraform.tf
@@ -10,11 +10,11 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "5.19.0"
+      version = "5.20.0"
     }
     pagerduty = {
       source  = "pagerduty/pagerduty"
-      version = "3.0.1"
+      version = "3.0.2"
     }
   }
 }